RIPE NCC has merged 1d7b6ca
* README: describe using static resource cache [skip ci] [a8dde72]
* Allow configuring static resource cache [c369e84]
* publish-source-code: find last published commit in <n> parents [3de1ba6]
* publish-source-code: enable verification by default [1e110ab]
* Update dependency org.sonarqube:org.sonarqube.gradle.plugin to v3.5.0.2730 [a9b5681]
* Update plugin com.google.cloud.tools.jib to v3.3.1 [f2c9bfe]
* Fix resource cache (tests) [1bd8ed8]
* Cleanup ROA prefix validation and improve error messages [740d586]
* Remove duplicate validation of ROA prefixes [2e9c808]
* Test when no update is needed for manifest or CRL [8064324]
* Test that ASPA and ROA entities get updated on manifest update [6d92063]
* Manage ASPA CMS objects when publishing [f10b546]
* Split CertificateManagementService [5bbd6c9]
* Fix merge error [3adb04f]
* Remove duplication [ed6b663]
* Move static factory methods suitable for unit tests to TestObjects [856c2fc]
* Remove `addOutgoingResourceCertificate` from `CertificateManagementService` [66d9e77]
* Remove unused `isManifestAndCrlUpdateNeeded` method [01d92c7]
* Remove unused certificate authority parameter when issuing EE certificate [5e52a0f]
* Remove unused import [bf5a987]
* Check name not unique constraint after adding new CA to JPA session [1e98202]
* Fix delete by key pair for ASPA and ROA entities [e63a55a]
* Update dependency com.google.code.gson:gson to v2.10 [3fed2b6]
* Update dependency org.easymock:easymock to v5.0.1 [b142191]
* Test when no update is needed for manifest or CRL [d473756]
* Test that ASPA and ROA entities get updated on manifest update [2c9fbe4]
* Manage ASPA CMS objects when publishing [719bd25]
* Split CertificateManagementService [a270417]
* Fix merge error [669a24a]
* Remove duplication [f5948e0]
* Move static factory methods suitable for unit tests to TestObjects [3894757]
* Remove `addOutgoingResourceCertificate` from `CertificateManagementService` [e23df33]
* Remove unused `isManifestAndCrlUpdateNeeded` method [15fe2bb]
* Remove unused certificate authority parameter when issuing EE certificate [3bc987c]
* Remove unused import [89d783d]
* Check name not unique constraint after adding new CA to JPA session [a718c8b]
* Update dependency io.sentry:sentry-bom to v6.6.0 [5678e33]
* Automatically subscribe all certificate authority event visitors [6c59927]
* Remove unneeded parameters for EE certificate subject name generation [0b2754c]
* Create and delete ASPA entities based on new/removed ASPA configuration [693edee]
* Support ASPA file names [e0d8c81]
* Ensure ROA entity is always removed when revoked [9c03267]
* Remove unused import [75ea690]
* Delete ASPA entities when issuing key pair is deleted [9afa64e]
* Add ASPA entity repository [20dafe3]
* Properly delete expired objects and ASPA entities on expiration [863d6e7]
* Make service transactional to prevent lazy initialization error [1bdac2a]
* KeyPairData is still used by Wicket UI [996f9df]
* CertificateAuthorityHistoryItem is still used by Wicket UI [8bc9dbd]
* CertificateAuthorityData is still used by the Wicket UI [ba1cf70]
* The ASPA customer ASN must not appear in the provider ASNs [c2cc21c]
* Update the ASPA configuration when certified resources change [4897bdb]
* ASPA must have at least one provider ASN configered per customer ASN [3f1fb48]
* Automatically subscribe all certificate authority event visitors [c63de31]
* Remove unneeded parameters for EE certificate subject name generation [6db73e2]
* Create and delete ASPA entities based on new/removed ASPA configuration [3aa1a41]
* Support ASPA file names [162e739]
* Ensure ROA entity is always removed when revoked [e55db9e]
* Remove unused import [d2925bb]
* CertificateAuthorityData is still used by the Wicket UI [650170d]
* Delete ASPA entities when issuing key pair is deleted [4107945]
* Add ASPA entity repository [ecf08ac]
* Properly delete expired objects and ASPA entities on expiration [a5449a0]
* Show actual error message when API keys fail to load [530763c]
* Close API keys file handle after reading [3408658]
* Update dependency org.springdoc:springdoc-openapi-ui to v1.6.12 [de6c088]
* Update dependency org.easymock:easymock to v5 [af460a6]
* Cleanup embedded validity period to fix Sonar maintainability warnings [72cf7ee]
* Remove unused java.io.Serializable implementations [792058c]
* Remove unused imports [aee294b]
* Use "no content" response when updating the ASPA configuration [0d362ea]
* Additional validation on updated ASPA configuration [742fb13]
* Ensure unique encoding of ASPA configuration for entity tag [e4f387f]
* chore(deps): update dependency io.sentry:sentry-bom to v6.5.0 [98fcaa4]
* Enable ASPA in all environments except production [caf9803]
* Slight cleanup (extract method) [6fd9ef0]
* Slightly more and initially faster command retries [9d7cd73]
* Command summary test for update ASPA configuration command [6b245bc]
* Fix some Sonar code smells [5e2e633]
* Error checking for input and missing or mismatched entity tag [f3257c3]
* Initial REST API to retrieve and update ASPA configuration for CA [ff131f3]
* Command handler to update ASPA configuration [09387e4]
* Enable readiness and liveness endpoints in actuator [0d729e7]
* gitlab: delete unused docker deployments [4e85854]
* AFI limit JavaDoc [ed267cd]
* Use enum to specify ASPA AFI Limit [8bd70eb]
* Use @Value instead of ValueObjectSupport [5d9940e]
* Use ASN persistence converter instead of converting manually [986506a]
* Ensure AFI limit can only be null, IPv4, or IPv6 [5192695]
* Allow multiple ASPA configurations per CA [f22c315]

+++ src/test/java/net/ripe/rpki/domain/archive/test/KeyPairDeletionServiceTest.java
@@ -1,43 +1,42 @@
 package net.ripe.rpki.domain.archive.test;
 
-import net.ripe.rpki.commons.crypto.util.KeyPairUtil;
+import net.ripe.rpki.domain.KeyPairEntity;
 import net.ripe.rpki.domain.ManagedCertificateAuthority;
 import net.ripe.rpki.domain.PublishedObjectRepository;
 import net.ripe.rpki.domain.ResourceCertificateRepository;
 import net.ripe.rpki.domain.TestObjects;
 import net.ripe.rpki.domain.archive.KeyPairDeletionService;
+import net.ripe.rpki.domain.aspa.AspaEntityRepository;
 import net.ripe.rpki.domain.crl.CrlEntityRepository;
 import net.ripe.rpki.domain.interca.CertificateRevocationResponse;
 import net.ripe.rpki.domain.manifest.ManifestEntityRepository;
 import net.ripe.rpki.domain.roa.RoaEntityRepository;
-import net.ripe.rpki.server.api.dto.KeyPairData;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import java.security.PublicKey;
 import java.util.Collections;
 import java.util.Optional;
+import java.util.function.Consumer;
 
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.matches;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class KeyPairDeletionServiceTest {
-    private static final String KEY_STORE_NAME = "keyStore";
-    private static final PublicKey PUBLIC_KEY = TestObjects.createTestKeyPair().getPublicKey();
-    private KeyPairDeletionService subject;
+    public static final KeyPairEntity KEY_PAIR = TestObjects.createTestKeyPair();
 
     @Mock
     private CrlEntityRepository crlEntityRepository;
     @Mock
     private ManifestEntityRepository manifestEntityRepository;
     @Mock
-    private RoaEntityRepository roaRepository;
+    private RoaEntityRepository roaEntityRepository;
+    @Mock
+    private AspaEntityRepository aspaEntityRepository;
     @Mock
     private ResourceCertificateRepository resourceCertificateRepository;
     @Mock
@@ -45,35 +44,27 @@ public class KeyPairDeletionServiceTest {
     @Mock
     private ManagedCertificateAuthority ca;
 
-    @Before
-    public void initKeyPairArchivingService() {
-        subject = new KeyPairDeletionService(crlEntityRepository,
-                manifestEntityRepository,
-                roaRepository,
-                resourceCertificateRepository,
-                publishedObjectRepository);
-    }
-
-    private Optional<KeyPairData> keyPairData(String keyStoreName) {
-        KeyPairData keyPairData = new KeyPairData(null, keyStoreName, null, null, null, null, null, null, false);
-        return Optional.of(keyPairData);
-    }
+    @InjectMocks
+    private KeyPairDeletionService subject;
 
     @Test
-    public void should_ask_the_ca_to_archive_revoked_keys() {
-        CertificateRevocationResponse revocationResponse = revocationResponse(PUBLIC_KEY);
-        when(ca.deleteRevokedKey(keyFrom(revocationResponse), crlEntityRepository, manifestEntityRepository, roaRepository, resourceCertificateRepository, publishedObjectRepository)).thenReturn(keyPairData(KEY_STORE_NAME));
+    @SuppressWarnings("unchecked")
+    public void should_ask_the_ca_to_delete_revoked_keys() {
+        CertificateRevocationResponse revocationResponse = new CertificateRevocationResponse("RIPE", KEY_PAIR.getPublicKey());
+
+        when(ca.deleteRevokedKey(any(), any())).then(invocation -> {
+            invocation.getArgument(1, Consumer.class).accept(KEY_PAIR);
+            return Optional.of(KEY_PAIR);
+        });
 
         subject.deleteRevokedKeysFromResponses(ca, Collections.singletonList(revocationResponse));
 
-        verify(ca).deleteRevokedKey(matches(keyFrom(revocationResponse)), any(CrlEntityRepository.class), any(ManifestEntityRepository.class), any(RoaEntityRepository.class), any(ResourceCertificateRepository.class), any(PublishedObjectRepository.class));
+        verify(publishedObjectRepository).withdrawAllForDeletedKeyPair(KEY_PAIR);
+        verify(crlEntityRepository).deleteByKeyPair(KEY_PAIR);
+        verify(manifestEntityRepository).deleteByKeyPairEntity(KEY_PAIR);
+        verify(roaEntityRepository).deleteByCertificateSigningKeyPair(KEY_PAIR);
+        verify(aspaEntityRepository).deleteByCertificateSigningKeyPair(KEY_PAIR);
+        verify(resourceCertificateRepository).deleteOutgoingCertificatesForRevokedKeyPair(KEY_PAIR);
     }
 
-    private String keyFrom(CertificateRevocationResponse revocationResponse) {
-        return KeyPairUtil.getEncodedKeyIdentifier(revocationResponse.getSubjectPublicKey());
-    }
-
-    private CertificateRevocationResponse revocationResponse(PublicKey publicKey) {
-        return new CertificateRevocationResponse("RIPE", publicKey);
-    }
 }
